app/src/test/java/nl/tudelft/b_b_w/controller/BlockControllerUnitTest.java
package nl.tudelft.b_b_w.ControllersUnitTest;

import android.content.res.Resources;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.robolectric.RobolectricTestRunner;
import org.robolectric.RuntimeEnvironment;
import org.robolectric.annotation.Config;

import java.util.ArrayList;
import java.util.List;

import nl.tudelft.b_b_w.BuildConfig;
import nl.tudelft.b_b_w.controller.BlockController;
import nl.tudelft.b_b_w.model.Block;
import nl.tudelft.b_b_w.model.BlockFactory;

import static android.R.id.list;
import static org.junit.Assert.assertEquals;

//import android.test.mock.MockContext;

/**
 * Example local unit test, which will execute on the development machine (host).
 *
 * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
 */
@RunWith(RobolectricTestRunner.class)
@Config(constants = BuildConfig.class,sdk= 21,  manifest = "src/main/AndroidManifest.xml")
public class BlockControllerUnitTest {

    /**
     * Attributes
     */
    private BlockController bc;
    private final String owner = "owner";
    private final int sequenceNumber = 1;
    private final String ownHash = "ownHash";
    private final String previousHashChain = "previousHashChain";
    private final String previousHashSender = "previousHashSender";
    private final String iban = "iban";
    private final String publicKey = "publicKey";
    private Block _block;
    private final String TYPE_BLOCK = "BLOCK";
    private final String TYPE_REVOKE = "REVOKE";

    /**
     * Initialize BlockController before every test
     * And initialize a dummy block _block
     */
    @Before
    public void setUp() {
        this.bc = new BlockController(RuntimeEnvironment.application);
        this._block = BlockFactory.getBlock(TYPE_BLOCK, owner, ownHash,
                previousHashChain, previousHashSender, publicKey, iban);
    }

    /**
     * Tests adding a block
     * @throws Exception RuntimeException
     */
    @Test
    public void testAddBlock() throws Exception {
        bc.addBlock(_block);
        List<Block> list = new ArrayList<>();
        list.add(_block);
        assertEquals(bc.getBlocks(owner), list);
    }


    /**
     * Tests returning the lastest block
     * @throws Exception RuntimeException
     */
    @Test
    public void testGetLastestBlock() throws Exception {
        Block expected = BlockFactory.getBlock(TYPE_BLOCK, owner, ownHash,
                previousHashChain, previousHashSender, publicKey, iban);
        bc.addBlock(_block);
        assertEquals(expected, bc.getLastestBlock(owner));
    }

    /**
     * Tests returning the lastestsequence number of chain
     * @throws Exception RuntimeException
     */
    @Test
    public void testGetLastestSeqNumber() throws Exception {
        String newOwner = owner+"2";
        Block newBlock = BlockFactory.getBlock(TYPE_BLOCK, newOwner, ownHash,
                previousHashChain, previousHashSender, publicKey,iban);
        bc.addBlock(_block);
        bc.addBlock(newBlock);


        assertEquals(1, bc.getLastestSeqNumber(owner));
    }




    /**
     * Tests adding two blocks
     * @throws Exception RuntimeException
     */
    @Test
    public void testAddBlock2() throws Exception {
        String newOwner = owner+"2";
        Block newBlock = BlockFactory.getBlock(TYPE_BLOCK, newOwner, ownHash,
                previousHashChain, previousHashSender, publicKey,iban);
        bc.addBlock(_block);
        bc.addBlock(newBlock);
        List<Block> list = new ArrayList<>();
        list.add(_block);
        list.add(newBlock);

        List<Block> newList = bc.getBlocks(owner);
        newList.addAll(bc.getBlocks(newOwner));

        assertEquals(newList, list);
}

    /**
     * Tests adding a duplicate block
     */
    @Test(expected=RuntimeException.class)
    public void testAddDupBlocks() {
        bc.addBlock(_block);
        bc.addBlock(_block);
    }

    /**
     * Tests adding an already revoked block
     */
    @Test(expected=RuntimeException.class)
    public void alreadyRevoked() {
        Block newBlock = BlockFactory.getBlock(TYPE_REVOKE, owner, ownHash,
                previousHashChain, previousHashSender, publicKey,iban);
        bc.addBlock(newBlock);
        bc.addBlock(_block);
    }

    /**
     * Tests filtering duplicates out of a list
     */
    @Test
    public void testEmptyList() {
        bc.addBlock(_block);
        Block newBlock = BlockFactory.getBlock(TYPE_BLOCK, owner, ownHash,
                previousHashChain, previousHashSender, publicKey,iban);
        bc.revokeBlock(newBlock);
        List<Block> list = new ArrayList<>();
        assertEquals(list, bc.getBlocks(owner));
    }

    /**
     * Test removeBlock if the specified revoked block has no match
     */
    @Test
    public void testRemoveWithNoMatch() throws Resources.NotFoundException{
        bc.addBlock(_block);
        Block blc2 = BlockFactory.getBlock(TYPE_BLOCK, owner+"2", ownHash,
                previousHashChain, previousHashSender, publicKey+"2", iban);
        bc.revokeBlock(blc2);
        List<Block> list = new ArrayList<>();
        list.add(_block);
        assertEquals(list, bc.getBlocks(owner));
    }

}